Don't do auto-fix in master service
As prededuct acquire an order lock, if the resource has been deleted, it
will do the auto-fix in prededuct, thus the auto-fix will requests the
same lock, but prededuct still acquire the lock, this will cause the
dead lock.

This will remove the auto-fix in master, let checker do this.

Change-Id: Ibe3f8e40e4f6d2c0f32366c720a6d1d40b84a064