Add batch management APIs for precharges
The precharge APIs currently support create, use, dispatch operation
for a single precharge and query operation without pagination. This
patch adds delete operation(not really delete it, just set the status
`deleted` to True), and supports create, dispatch, delete a batch of
precharges. Also adds a column `operator_id` to record the operator
when someone creates a precharge and supports pagination for query a
list of precharges.

Implements task: BILL-67
Change-Id: Ie1e81a3af6502ddb5f00d3de7d9abb9ca033b423