Fix the bug what the salesperson is null when the sales_id exists
* When use the api with the param limit and the limit is 10,then
the API collects 10 user_ids to a list, but the sales_id is not
in the user_ids list, and use the keystone API to get the user's
info. So we can't get the salesperson from the list if the sales_id
is not in the 10 user_ids.
* I fix it, When it collects the user_id, it also collects the
sales_id of the user.Then the list of ids is enough to use.It can
get the salesperson info from the keystone.

implements task: BILL-102

Change-Id: I82c8a9af0ba57f90d85bd6e4bb7ae872c7b9db53